# swagger.yaml
openapi: 3.0.0
info:
    title: Playgram API
    version: 1.0.0
    description: API documentation for a social media application.
servers:
    - url: http://localhost:3000/
      description: Local Development Server
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: token
    schemas:
        User:
            type: object
            properties:
                id:
                    type: integer
                    format: int32
                    description: Unique identifier for the user.
                username:
                    type: string
                    description: Unique username of the user.
                email:
                    type: string
                    format: email
                    description: Unique email address of the user.
                passwordHash:
                    type: string
                    description: Hashed password of the user (Not exposed).
                fullName:
                    type: string
                    description: Full name of the user.
                bio:
                    type: string
                    nullable: true
                    description: Biography of the user.
                profilePicture:
                    type: string
                    nullable: true
                    description: URL to the user's profile picture.
                followerCount:
                    type: integer
                    format: int32
                    default: 0
                    description: Number of followers the user has.
                followingCount:
                    type: integer
                    format: int32
                    default: 0
                    description: Number of users the user is following.
                postsCount:
                    type: integer
                    format: int32
                    default: 0
                    description: Number of posts the user has created.
                isPrivate:
                    type: boolean
                    default: false
                    description: Indicates if the user's profile is private.
                createdAt:
                    type: string
                    format: date-time
                    description: Date and time when the user account was created.
                updatedAt:
                    type: string
                    format: date-time
                    description: Date and time when the user account was last updated.
...
paths:
    /users:
        get:
            summary: Get all users
            tags:
                - Users
            responses:
                200:
                    description: A list of users.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/User'
                500:
                    description: Failed to fetch users.
...
    /users/{username}:
        get:
            summary: Get user by username
            tags:
                - Users
            parameters:
                - in: path
                  name: username
                  schema:
                      type: string
                  required: true
                  description: Username of the user to retrieve.
            responses:
                200:
                    description: User data.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                404:
                    description: User not found.
                500:
                    description: Failed to fetch user.
    /users/{id}:
        put:
            summary: Update user profile
            tags:
                - Users
            security:
                - cookieAuth: []
            parameters:
                - in: path
                  name: id
                  schema:
                      type: integer
                  required: true
                  description: Numeric ID of the user to update.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                username:
                                    type: string
                                email:
                                    type: string
                                    format: email
                                fullName:
                                    type: string
                                bio:
                                    type: string
                                profilePicture:
                                    type: string
                                isPrivate:
                                    type: boolean
            responses:
                200:
                    description: Profile updated successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                    user:
                                        $ref: '#/components/schemas/User'
                403:
                    description: You are not allowed to edit this profile.
                400:
                    description: Failed to update profile.
        delete:
            summary: Delete a user
            tags:
                - Users
            parameters:
                - in: path
                  name: id
                  schema:
                      type: integer
                  required: true
                  description: Numeric ID of the user to delete.
            responses:
                204:
                    description: User deleted successfully.
                400:
                    description: Failed to delete user.
...